Prepare for weather API addition
First is a new DRY_RUN mode (kind of surprised I hadn't put that in now
that I think about it). Then it's moving the locations to a file with a
JSON blob (I was lazy and needed objects, this can be changed if the
need arises). Then it's assigning temperature limits (in Celsius) to the
locations so that they're picked. The current temperature is a stub
until the API call is put in.