Optimize algorithm for better location spread
The old algorithm had a minor flaw where outside of the 5 week window a
location could be chosen again and again, and potentially be chosen when
it wouldn't be ideal (eg a wet-weather location would get chosen on a
nice forecast, because "random"). This meant the bot would cluster
around a set of locations that had a sweet spot of temperature ranges
and rainy-weather protection. It also meant there was a theoretical
edge-case where the bot could fail to pick a location if there wasn't
enough locations at the temperature extremes.

The new algorithm should address this, where it goes through the
*entire* list of previous CoffeeOutside locations, and picks the
location that was least recently used (with a random selection if
there's a few more than one new location). This pseudo-determinism would
have been undesirable when there was less than 10 locations (oh gosh,
the bot was so modest in 2016!), but now there's over 60 locations with
more on the way, so this seems like a good way to see lots of
parks/cafes which might only be desirable in narrow bands of
temperature, and get more use out of all the locations that have been
added over the years.