Don't use twitter handles for locations
As Twitter seems to be having some problems as of late, it seems that
now is a good time to assign the proper names to locations that
previously had been identified by their Twitter handles. We'll still use
the handles on the Twitter dispatcher, but everywhere else we'll use the
proper name.