Breaking change: Twitter support has been removed
Not a surprise to anyone who's checked the @coffeeoutside bot in the
past week. I'm not paying a nickel to generate content for that
platform, let alone $100/month. While I could have made Twitter support
optional, I don't know anyone else currently using the bot, and
unexercised code tends to rot. If you do use the bot and plan on paying,
ping me (because the pain of maintaining a fork sucks).